Java doc + evaluer dans SolutionMock.java
