+++ src/main/model/generic/Solution.java
@@ -79,7 +79,7 @@ public abstract class Solution {
 	private List<Double> getValueObjective() {
 		return valuesObjectives;
 	}
-	private int[] getValuesVariables() {
+	public int[] getValuesVariables() {
 		return valueVariables;
 	}
 	public int getNbVariables() {
