+++ src/test/model/algorithmes/fourmis/AlgoFourmisTest.java
@@ -11,6 +11,9 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+/**
+ * @author Robin LECORVAISIER
+ */
 class AlgoFourmisTest {
