--- src/test/model/algorithmes/fourmis/FourmisTest.java
@@ -1,5 +1,0 @@
-package test.model.algorithmes.fourmis;
-
-class FourmisTest {
-
-}
